+++ inc/common/PID/PID.cpp
@@ -12,7 +12,7 @@
 
 using namespace digitalcave;
 
-PID::PID(double kp, double ki, double kd, uint8_t direction, uint32_t time) :
+PID::PID(float kp, float ki, float kd, uint8_t direction, uint32_t time) :
 	lastTime(time),
 	integratedError(0),
 	lastMeasured(0)
@@ -22,35 +22,35 @@ PID::PID(double kp, double ki, double kd, uint8_t direction, uint32_t time) :
 	setTunings(kp, ki, kd);
 }
 
-double PID::compute(double setPoint, double measured, uint32_t time){
+float PID::compute(float setPoint, float measured, uint32_t time){
 	uint32_t currentPeriod = time - lastTime;
 	if (currentPeriod == 0) currentPeriod = 1;	//Calling this more frequently than once per ms is not recommended...
-	double periodSec = currentPeriod / 1000.0;	//Period in seconds; used to normalize ki / kd values across different periods
-	
+	float periodSec = currentPeriod / 1000.0;	//Period in seconds; used to normalize ki / kd values across different periods
+
 	//Compute all the working error variables
-	double error = setPoint - measured;
+	float error = setPoint - measured;
 	integratedError += (ki * periodSec * error);
 
 	if (integratedError > outMax) integratedError = outMax;
 	else if (integratedError < outMin) integratedError = outMin;
-	
-	double derivative = (measured - lastMeasured);
+
+	float derivative = (measured - lastMeasured);
 
 	// Compute PID Output
-	double result = kp * error + integratedError - kd / periodSec * derivative;
-	
+	float result = kp * error + integratedError - kd / periodSec * derivative;
+
 	if (result > outMax) result = outMax;
 	else if(result < outMin) result = outMin;
-	
+
 	//Remember some variables for next time
 	lastMeasured = measured;
 	lastTime = time;
-	
+
 	return result;
 }
 
 
-void PID::setTunings(double kp, double ki, double kd){
+void PID::setTunings(float kp, float ki, float kd){
 	if (kp < 0 || ki < 0 || kd < 0) return;
 
 	if(direction == DIRECTION_REVERSE){
@@ -64,8 +64,8 @@ void PID::setTunings(double kp, double ki, double kd){
 		this->kd = kd;
 	}
 }
-	
-void PID::setOutputLimits(double min, double max){
+
+void PID::setOutputLimits(float min, float max){
 	if (min >= max) return;
 
 	outMin = min;
@@ -87,24 +87,24 @@ void PID::reset(uint32_t time){
 	lastTime = time;
 }
 
-double PID::getKp() {
+float PID::getKp() {
 	return kp;
 }
 
-double PID::getKi() {
+float PID::getKi() {
 	return ki;
 }
 
-double PID::getKd() {
+float PID::getKd() {
 	return kd;
 }
 
-void PID::setKp(double Kp){
+void PID::setKp(float Kp){
 	kp = Kp;
 }
-void PID::setKi(double Ki){
+void PID::setKi(float Ki){
 	ki = Ki;
 }
-void PID::setKd(double Kd){
+void PID::setKd(float Kd){
 	kd = Kd;
 }
